fix so that no error is given in init display
